Terminal appearance

+++ .input/sim/simGADGET.sh
@@ -15,7 +15,7 @@ status_file "STARTING.tmp"
 
 # check for ATTPCROOT installed in same directory, download if not
 if [ ! -f $automation_dir"ATTPCROOTv2/env_fishtank.sh" ]; then
-    status_file "DOWNLOADING.tmp"
+    status_file "CLONING_ATTPCROOT.tmp"
     git clone https://github.com/ATTPC/ATTPCROOTv2
     cd ATTPCROOTv2
     git checkout 40699a2
@@ -30,14 +30,19 @@ source $attpcroot_dir"env_fishtank.sh"
 
 if [ ! -f $attpcroot_dir"build/Makefile" ]; then
     # steps pulled from ATTPCROOTv2 installation wiki page
-    status_file "INSTALLING.tmp"
+    status_file "CMAKING_ATTPCROOT.tmp"
     cd $attpcroot_dir
     source env_fishtank.sh
     mkdir build
     cd build
     cmake -DCMAKE_PREFIX_PATH=/mnt/simulations/attpcroot/fair_install_18.6/ ../
+    cd $automation_dir
+    status_file "MAKING_ATTPCROOT.tmp"
+    cd $attpcroot_dir"build/"
     make install -j4
     source config.sh
+    cd $automation_dir
+    status_file "GENERATING_GEOMETRY.tmp"
     cd $attpcroot_dir"geometry/"
     nohup root -b -l GADGET_II.C
     pid1=$!
@@ -66,13 +71,19 @@ while true; do
         exit 0
     fi
     
+    if [ -f ../STOP.tmp ]; then
+        status_file "STOPPED.tmp"
+        exit 0
+    fi
+    
     # build ATTPCROOT and run simulation
+    status_file "REBUILDING.tmp"
     make -C $attpcroot_dir"build/" -j8
     
     if [ -f "geo.temp" ]; then
         # update geometry if geo.temp exists
         rm "geo.temp"
-        status_file "GEOMETRY.tmp"
+        status_file "UPDATING_GEOMETRY.tmp"
         cd $attpcroot_dir"geometry/"
         nohup root -b -l GADGET_II.C
         pid1=$!
@@ -80,14 +91,14 @@ while true; do
         cd $automation_dir
     fi
     
-    status_file "Sim.tmp"
+    status_file "SIMULATING.tmp"
     cd $attpcroot_dir"macro/Simulation/Charge_Dispersion/"
     echo "Running simulation (Mg20_test_sim_pag.C)"
     nohup root -b -l Mg20_test_sim_pag.C
     pid1=$!
     wait $pid1
     cd $automation_dir
-    status_file "RUNDIGI.tmp"
+    status_file "DIGITIZATION.tmp"
     cd $attpcroot_dir"macro/Simulation/Charge_Dispersion/"
     nohup root -b -l rundigi_sim_CD.C
     pid2=$!
